Feat: 배포파일 추가

See merge request s09-webmobile1-sub2/S09P12E202!46

+++ backend/src/main/java/com/ssafish/web/dto/MsgRequest.java
@@ -8,6 +8,6 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MsgRequest {
-    int roomId;
+    Long roomId;
     String content;
 }
